+++ app/src/main/java/com/example/northwind/History.java
@@ -50,7 +50,7 @@ public class History extends AppCompatActivity{
         Realm.init(this);
         App app=new App(new AppConfiguration.Builder("northwind-noimz").build());
         ArrayList<String> arrayList=new ArrayList<>();
-        ArrayList<Integer> arrayList1=new ArrayList();
+        ArrayList<Double> arrayList1=new ArrayList();
         ArrayList<String> userInfo=new ArrayList<>();
         ArrayList<String> action=new ArrayList<>();
 
@@ -84,7 +84,7 @@ public class History extends AppCompatActivity{
                                 if(currentDoc.getString("food")!=null){
 //                                        Log.d("aaa",currentDoc.toString());
                                     arrayList.add(currentDoc.getString("food"));
-                                    arrayList1.add(currentDoc.getInteger("price"));
+                                    arrayList1.add(currentDoc.getDouble("price"));
                                 }
                             }
                             try{
